handlers: rename filtered to filter
