+++ .vscode/settings.json
@@ -2,7 +2,7 @@
 	"files.exclude": {
 		"handlers/cblog": true,
 		"handlers/discard": true,
-		"handlers/filtered": true,
+		"handlers/filter": true,
 		"handlers/logrus": true,
 		"handlers/zap": true,
 		"handlers/zerolog": true
