+++ Makefile
@@ -16,7 +16,7 @@ V = 0
 Q = $(if $(filter 1,$V),,@)
 M = $(shell if [ "$$(tput colors 2> /dev/null || echo 0)" -ge 8 ]; then printf "\033[34;1m▶\033[0m"; else printf "▶"; fi)
 
-HANDLERS = cblog discard filtered logrus zap zerolog
+HANDLERS = cblog discard filter logrus zap zerolog
 
 TMPDIR ?= .tmp
